+++ CHANGELOG.md
@@ -5,5 +5,5 @@
 ## [v1.0.0] / 2026-mm-dd
 ### Initial release
 - Add basic AIS stream subscription to OpenWaters.io
-- Select search position (on boat or on cursor)
-- Select search box size up to 10°x10°(centered on selected search position)
+- Search position on boat (follow mode) or on cursor (manual selection)
+- Select search box size up to 10°x10° (centered on selected search position)